+++ .gitignore
@@ -2,6 +2,9 @@
 __pycache__/
 *.py[cod]
 *$py.class
+secret.sh
+.secret
+
 
 # C extensions
 *.so
